trying to solve the data reception
the data comes in in blocks for unknown reasons or are read in that way
from the input stream. every time the input stream is read an intent is
created and the data is broadcasted to LocalBroadcastManager that will
then displays the text in the incomingMessage textfield. to display the
data correctly, these incoming messages have to be appended to each
other. This poses a problem, when the data needs to be parsed or
identified. Trying to do that now but I think I will need help from
somebody more knowledgeable.